requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
redis==5.0.1
langchain==0.1.1
langchain-community==0.0.13
jinaai==0.2.10
qdrant-client==1.7.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.26.0
google-generativeai==0.3.2
feedparser==6.0.10
fastapi-socketio==0.0.10
python-multipart==0.0.6
httpx==0.25.1
aiofiles==23.2.1